+++ src/components/app/AppReloadPrompt.vue
@@ -1,7 +1,19 @@
 <template>
   <div
     v-if="needRefresh"
-    class="absolute p-5 absolute z-50 bottom-0 right-0 bg-gray-700 m-5 max-w-xs shadow-xl rounded-xl"
+    class="
+      absolute
+      p-5
+      absolute
+      z-50
+      bottom-0
+      right-0
+      bg-gray-700
+      m-5
+      max-w-xs
+      shadow-xl
+      rounded-xl
+    "
   >
     <p class="leading-tight">New content available, click on reload button to update.</p>
     <div class="space-x-2 flex mt-4">