+++ src/components/app/AppSidebar.vue
@@ -1,39 +1,53 @@
 <template>
   <div
-    class="absolute z-50 lg:relative w-full lg:w-64 bg-black bg-opacity-25 h-screen"
+    class="fixed top-0 left-0 w-full lg:w-64 bg-black bg-opacity-25 h-screen"
     :class="[store.state.showSidebar ? 'block' : 'hidden lg:block']"
+    style="z-index: 999"
     @click.self="closeSidebar"
   >
     <div
       class="bg-light h-full p-5 overflow-auto scroll w-64 bg-opacity-100 shadow-xl lg:shadow-none"
     >
       <div class="mb-6 library">
-        <p class="mb-3 text-lighter">Library</p>
         <list-ui class="space-y-1">
-          <list-item-ui
-            v-for="filter in filters"
-            :key="filter.id"
-            small
-            class="cursor-pointer"
-            :active="activeFilter === filter.id"
-            @click="filterBy(filter.id)"
-          >
+          <list-item-ui small tag="router-link" :active="$route.name === 'explore'" to="/explore">
             <template #prepend>
-              <icon-ui :name="filter.icon"></icon-ui>
+              <v-mdi name="mdi-compass-outline"></v-mdi>
             </template>
-            {{ filter.name }}
+            Explore
           </list-item-ui>
+          <list-item-ui v-if="!isInHome" small tag="router-link" to="/">
+            <template #prepend>
+              <v-mdi name="mdi-archive-outline"></v-mdi>
+            </template>
+            My Snippets
+          </list-item-ui>
+          <template v-else>
+            <list-item-ui
+              v-for="filter in filters"
+              :key="filter.id"
+              small
+              class="cursor-pointer"
+              :active="activeFilter === filter.id"
+              @click="filterBy(filter.id)"
+            >
+              <template #prepend>
+                <v-mdi :name="filter.icon"></v-mdi>
+              </template>
+              {{ filter.name }}
+            </list-item-ui>
+          </template>
         </list-ui>
       </div>
-      <div class="folders">
+      <div v-if="isInHome" class="folders">
         <div class="flex items-center text-lighter justify-between mb-3">
           <p>Folders</p>
-          <icon-ui
+          <v-mdi
             v-tooltip="'Add folder'"
             class="cursor-pointer"
-            name="plus"
+            name="mdi-plus"
             @click="addFolder"
-          ></icon-ui>
+          ></v-mdi>
         </div>
         <sidebar-folders
           v-bind="{ activeFilter }"
@@ -46,6 +60,7 @@
 <script>
 import { computed } from 'vue';
 import { useStore } from 'vuex';
+import { useRoute } from 'vue-router';
 import { useDialog } from '~/composable';
 import { Folder } from '~/models';
 import SidebarFolders from './sidebar/SidebarFolders.vue';
@@ -54,14 +69,16 @@ export default {
   components: { SidebarFolders },
   setup() {
     const filters = [
-      { id: 'all', name: 'All Snippets', icon: 'archive' },
-      { id: 'starred', name: 'Starred', icon: 'star' },
+      { id: 'all', name: 'All Snippets', icon: 'mdi-archive-outline' },
+      { id: 'starred', name: 'Starred', icon: 'mdi-star-outline' },
     ];
 
     const store = useStore();
+    const route = useRoute();
     const dialog = useDialog();
 
     const activeFilter = computed(() => store.state.filterBy);
+    const isInHome = computed(() => ['home', 'view'].includes(route.name));
 
     function filterBy(name) {
       store.commit('updateState', { key: 'filterBy', value: name });
@@ -97,6 +114,7 @@ export default {
     return {
       store,
       filters,
+      isInHome,
       filterBy,
       addFolder,
       closeSidebar,