+++ src/components/pages/home/HomeSnippets.vue
@@ -8,7 +8,7 @@
           class="rounded-r-none border-r align-bottom"
           @click="sort.type = sort.type === 'asc' ? 'desc' : 'asc'"
         >
-          <icon-ui :name="sort.type === 'asc' ? 'sortAscending' : 'sortDescending'"></icon-ui>
+          <v-mdi :name="sort.type === 'asc' ? 'mdiSortAscending' : 'mdiSortDescending'"></v-mdi>
         </button-ui>
         <select-ui v-model="sort.by" class="rounded-l-none">
           <option v-for="option in sortOptions" :key="option.value" :value="option.value">
@@ -16,15 +16,32 @@
           </option>
         </select-ui>
       </div>
-      <button-ui
-        v-tooltip="'Add snippet'"
-        icon
-        variant="primary"
-        :disabled="activeFilter === 'starred' || activeFilter === 'all'"
-        @click="addFile"
-      >
-        <icon-ui name="plus"></icon-ui>
-      </button-ui>
+      <popover-ui>
+        <button-ui icon variant="primary" :disabled="activeFilter === 'starred'">
+          <v-mdi name="mdi-plus"></v-mdi>
+        </button-ui>
+        <template #popover>
+          <list-ui class="space-y-1 w-44">
+            <list-item-ui v-close-popover small class="cursor-pointer" @click="addFile">
+              <template #prepend>
+                <v-mdi name="mdi-file-outline"></v-mdi>
+              </template>
+              <span>Add snippet</span>
+            </list-item-ui>
+            <list-item-ui
+              v-close-popover
+              small
+              class="cursor-pointer"
+              @click="emitter.emit('import-gist')"
+            >
+              <template #prepend>
+                <v-mdi name="mdi-github"></v-mdi>
+              </template>
+              <span>Import gists</span>
+            </list-item-ui>
+          </list-ui>
+        </template>
+      </popover-ui>
     </div>
     <div v-if="!isRetrieved" class="skeletons">
       <div v-for="i in 5" :key="i" class="border-b py-4 px-5 w-full bg-input animate-pulse">
@@ -39,6 +56,7 @@
 <script>
 import { computed, shallowReactive } from 'vue';
 import { useStore } from 'vuex';
+import emitter from 'tiny-emitter/instance';
 import { File } from '~/models';
 import SnippetList from './snippet/SnippetList.vue';
 
@@ -82,9 +100,10 @@ export default {
       File.$update({
         data: {
           name: 'untitled snippet',
-          folderId: store.state.filterBy,
+          folderId: store.state.filterBy === 'all' ? '' : store.state.filterBy,
           isNew: true,
           isEdited: true,
+          createdAt: Date.now(),
         },
       });
       store.commit('updateState', {
@@ -96,6 +115,7 @@ export default {
     return {
       sort,
       addFile,
+      emitter,
       snippets,
       sortOptions,
       isRetrieved,