+++ src/components/pages/home/gists/GistsList.vue
@@ -8,12 +8,12 @@
         icon
         @click="toggleSelectFiles(state.selectedFiles.length >= files.length)"
       >
-        <icon-ui :name="files.length === state.selectedFiles ? 'close' : 'check'"></icon-ui>
+        <v-mdi :name="files.length === state.selectedFiles ? 'mdi-close' : 'mdi-check'"></v-mdi>
       </button-ui>
       <input-ui
         v-model="state.query"
         placeholder="Search file"
-        prepend-icon="search"
+        prepend-icon="mdi-magnify"
         class="w-full"
       ></input-ui>
     </div>
@@ -41,12 +41,7 @@
           </option>
         </select-ui>
         <div class="flex-grow"></div>
-        <button-ui
-          variant="primary"
-          :disabled="state.selectedFolder === ''"
-          :loading="state.loadingImport"
-          @click="importGists"
-        >
+        <button-ui variant="primary" :loading="state.loadingImport" @click="importGists">
           Import ({{ state.selectedFiles.length }})
         </button-ui>
       </div>
@@ -109,33 +104,46 @@ export default {
     async function importGists() {
       state.loadingImport = true;
 
-      for (const id of state.selectedFiles) {
-        const { content, language, raw_url, filename } =
-          props.files.find((file) => file.id === id) || {};
-        const file = {
-          name: filename,
-          isEdited: true,
-          isNew: true,
-          folderId: state.selectedFolder,
-          language: language.toLowerCase(),
-          code: content,
-        };
+      const promises = state.selectedFiles.map(async (id) => {
+        try {
+          const { content, language, raw_url, filename } =
+            props.files.find((file) => file.id === id) || {};
+          const file = {
+            name: filename,
+            isEdited: true,
+            isNew: true,
+            folderId: state.selectedFolder,
+            language: language.toLowerCase(),
+            code: content,
+            createdAt: Date.now(),
+          };
 
-        if (typeof content === 'undefined') {
-          const response = await fetch(raw_url);
-          const code = await response.text();
+          if (typeof content === 'undefined') {
+            const response = await fetch(raw_url);
+            const code = await response.text();
 
-          file.code = code;
+            file.code = code;
+          }
+
+          return file;
+        } catch (error) {
+          return error;
         }
+      });
+      const snippets = await Promise.allSettled(promises);
 
-        File.$update({
-          data: file,
-        });
-        store.commit('updateState', {
-          key: 'isDataChanged',
-          value: true,
-        });
-      }
+      snippets.forEach(({ status, value }) => {
+        if (status === 'fulfilled') {
+          File.$update({
+            data: value,
+          });
+        }
+      });
+
+      store.commit('updateState', {
+        key: 'isDataChanged',
+        value: true,
+      });
 
       state.loadingImport = false;
       state.selectedFiles = [];