+++ src/components/pages/snippet/SnippetButtonsGroup.vue
@@ -1,35 +1,29 @@
 <template>
   <button-group-ui class="divide-x">
-    <popover-ui v-if="!isLocalFile">
+    <popover-ui v-if="!isFromFork" class="button-ui">
       <button-ui v-tooltip.group="'Fork snippet'" icon>
-        <icon-ui name="mdiSourceFork"></icon-ui>
+        <v-mdi name="mdiSourceFork"></v-mdi>
       </button-ui>
       <template #popover>
-        <select-ui v-model="selectedFolder" class="w-full">
+        <select-ui v-model="selectedFolder" class="w-full" placeholder="Select folder">
           <option v-for="folder in folders" :key="folder.id" :value="folder.id">
             {{ folder.name }}
           </option>
         </select-ui>
-        <button-ui
-          v-close-popover
-          block
-          variant="primary"
-          class="mt-4"
-          :disabled="selectedFolder === ''"
-          @click="forkSnippet"
-        >
+        <button-ui v-close-popover block variant="primary" class="mt-4" @click="forkSnippet">
           Fork Snippet
         </button-ui>
       </template>
     </popover-ui>
     <button-ui v-tooltip.group="'Copy code'" icon @click="copyCode">
-      <icon-ui name="clipboardCopy"></icon-ui>
+      <v-mdi name="mdi-clipboard-outline"></v-mdi>
     </button-ui>
   </button-group-ui>
 </template>
 <script>
-import { onMounted, ref } from 'vue';
+import { onMounted, ref, computed } from 'vue';
 import { useToast } from 'vue-toastification';
+import { nanoid } from 'nanoid';
 import { useGroupTooltip } from '~/composable';
 import { Folder, File } from '~/models';
 import { copyToClipboard } from '~/utils/helper';
@@ -40,14 +34,18 @@ export default {
       type: Object,
       default: () => ({}),
     },
-    isLocalFile: Boolean,
   },
-  emits: ['fork'],
   setup(props, { emit }) {
     const toast = useToast();
 
     const folders = Folder.all();
     const selectedFolder = ref(folders[0]?.id || '');
+
+    const isFromFork = computed(() =>
+      File.query()
+        .where((file) => file.id.includes(props.file.id))
+        .exists()
+    );
 
     function copyCode() {
       copyToClipboard(props.file.code);
@@ -59,27 +57,24 @@ export default {
     function forkSnippet() {
       const copyFile = { ...props.file };
 
-      delete copyFile.id;
-
-      File.$create({
+      File.$update({
         data: {
           ...copyFile,
+          id: `${nanoid()}___${copyFile.id.split('___')[0]}`,
           createdAt: Date.now(),
+          folderId: selectedFolder.value,
           isEdited: true,
           isNew: true,
         },
-      }).then((file) => {
-        emit('fork', file);
       });
     }
 
-    onMounted(() => {
-      useGroupTooltip();
-    });
+    onMounted(useGroupTooltip);
 
     return {
       folders,
       copyCode,
+      isFromFork,
       forkSnippet,
       selectedFolder,
     };