+++ src/router/index.js
@@ -1,9 +1,11 @@
 import { createRouter, createWebHistory } from 'vue-router';
+import store from '../store';
 import Home from '../views/Home/Index.vue';
 import View from '../views/Home/View.vue';
 import Snippet from '../views/Snippet.vue';
 import Auth from '../views/Auth.vue';
 import PageNotFound from '../views/404.vue';
+import Explore from '../views/Explore.vue';
 
 const routes = [
   {
@@ -22,27 +24,40 @@ const routes = [
     path: '/snippet/:fileId',
     name: 'snippet',
     component: Snippet,
+    meta: {
+      hideSidebar: true,
+    },
   },
   {
     path: '/auth',
     name: 'auth',
     component: Auth,
+    meta: {
+      hideSidebar: true,
+    },
   },
   {
-    path: '/404',
-    name: 'not-found',
-    component: PageNotFound,
+    path: '/explore',
+    name: 'explore',
+    component: Explore,
   },
   {
     path: '/:pathMatch(.*)*',
     name: 'not-found',
     component: PageNotFound,
+    meta: {
+      hideSidebar: true,
+    },
   },
 ];
 
 const router = createRouter({
   history: createWebHistory(),
   routes,
+});
+
+router.afterEach(() => {
+  store.commit('updateState', { key: 'historyState', value: history.state });
 });
 
 export default router;